requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
requests
pydantic
aiohttp
pydantic-settings
httpx
pillow
google-generativeai
pydantic[email]